adding of lexis_analys

+++ frontend/tree/graph_lib/graph_functions.cpp
@@ -29,7 +29,7 @@ void init_graph (dump_graph_t* graph_dump_set, const char* dot_dir)
 
 //--------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-void print_int_node (dump_graph_t* graph_dump_set, int* node_address, struct node_t nodes, int* right, int* left, int value)
+void print_int_node (dump_graph_t* graph_dump_set, double* node_address, struct node_t nodes, double* right, double* left, int value)
 {
     if (graph_dump_set->node_capacity <= graph_dump_set->node_size + 1) resize_struct (graph_dump_set);
 
@@ -38,7 +38,7 @@ void print_int_node (dump_graph_t* graph_dump_set, int* node_address, struct nod
     graph_dump_set->node_size++;
 }
 
-void print_char_node (dump_graph_t* graph_dump_set, int* node_address, struct node_t nodes, int* right, int* left, char value)
+void print_char_node (dump_graph_t* graph_dump_set, double* node_address, struct node_t nodes, double* right, double* left, char value)
 {
     if (graph_dump_set->node_capacity <= graph_dump_set->node_size + 1) resize_struct (graph_dump_set);
 
@@ -47,7 +47,7 @@ void print_char_node (dump_graph_t* graph_dump_set, int* node_address, struct no
     graph_dump_set->node_size++;
 }
 
-void print_str_node (dump_graph_t* graph_dump_set, int* node_address, struct node_t nodes, int* right, int* left, const char* value)
+void print_str_node (dump_graph_t* graph_dump_set, double* node_address, struct node_t nodes, double* right, double* left, const char* value)
 {
     if (graph_dump_set->node_capacity <= graph_dump_set->node_size + 1) resize_struct (graph_dump_set);
 
@@ -58,7 +58,7 @@ void print_str_node (dump_graph_t* graph_dump_set, int* node_address, struct nod
 
 //--------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-void make_edge (dump_graph_t* graph_dump_set, int* node_from, int* node_to, struct edge_t edges)
+void make_edge (dump_graph_t* graph_dump_set, double* node_from, double* node_to, struct edge_t edges)
 {
     if (graph_dump_set->edge_capacity <= graph_dump_set->edge_size + 1) resize_struct (graph_dump_set);
